+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Added
+- (metrics-strongswan.rb, metrics-strongswan-connection.rb) Fixed bug with regexp (@yuri-zubov)
+
 
 ## [2.1.0] - 2018-03-17
 ### Added